backend: remove getBackendIO
