+++ pkg/proxy/backend/common_test.go
@@ -20,8 +20,8 @@ import (
 	"testing"
 
 	"github.com/pingcap/TiProxy/lib/util/security"
+	"github.com/pingcap/TiProxy/lib/util/waitgroup"
 	pnet "github.com/pingcap/TiProxy/pkg/proxy/net"
-	"github.com/pingcap/tidb/util"
 	"github.com/stretchr/testify/require"
 )
 
@@ -56,7 +56,7 @@ func newTCPConnSuite(t *testing.T) *tcpConnSuite {
 }
 
 func (tc *tcpConnSuite) newConn(t *testing.T, withBackend bool) func() {
-	var wg util.WaitGroupWrapper
+	var wg waitgroup.WaitGroup
 	if withBackend {
 		wg.Run(func() {
 			conn, err := tc.backendListener.Accept()
@@ -92,7 +92,7 @@ func (tc *tcpConnSuite) newConn(t *testing.T, withBackend bool) func() {
 }
 
 func (tc *tcpConnSuite) run(clientRunner, backendRunner func(*pnet.PacketIO) error, proxyRunner func(*pnet.PacketIO, *pnet.PacketIO) error) (cerr, berr, perr error) {
-	var wg util.WaitGroupWrapper
+	var wg waitgroup.WaitGroup
 	if clientRunner != nil {
 		wg.Run(func() {
 			cerr = clientRunner(tc.clientIO)
